+++ packages/vendure-plugin-goedgepickt/src/goedgepickt.plugin.ts
@@ -1,4 +1,5 @@
 import {
+  CustomOrderFields,
   PluginCommonModule,
   RuntimeVendureConfig,
   VendurePlugin,
@@ -13,15 +14,13 @@ import {
 } from './api/goedgepickt.resolver';
 import { PLUGIN_INIT_OPTIONS } from './constants';
 import { schema } from './api/schema.graphql';
-import { GoedgepicktConfigEntity } from './api/goedgepickt-config.entity';
 import path from 'path';
 import { AdminUiExtension } from '@vendure/ui-devkit/compiler';
-import { customFields } from './api/custom-fields';
+import { channelCustomFields, orderCustomFields } from './custom-fields';
 
 @VendurePlugin({
   imports: [PluginCommonModule],
   controllers: [GoedgepicktController],
-  entities: [GoedgepicktConfigEntity],
   providers: [
     GoedgepicktService,
     {
@@ -36,7 +35,8 @@ import { customFields } from './api/custom-fields';
   configuration: (config: RuntimeVendureConfig) => {
     config.shippingOptions.fulfillmentHandlers.push(goedgepicktHandler);
     config.authOptions.customPermissions.push(goedgepicktPermission);
-    config.customFields.Order.push(...customFields.Order!);
+    config.customFields.Order.push(...orderCustomFields);
+    config.customFields.Channel.push(...channelCustomFields);
 
     return config;
   },
@@ -56,12 +56,6 @@ export class GoedgepicktPlugin {
   static ui: AdminUiExtension = {
     extensionPath: path.join(__dirname, 'ui'),
     ngModules: [
-      {
-        type: 'lazy',
-        route: 'goedgepickt',
-        ngModuleFileName: 'goedgepickt.module.ts',
-        ngModuleName: 'GoedgepicktModule',
-      },
       {
         type: 'shared',
         ngModuleFileName: 'goedgepickt-nav.module.ts',