+++ packages/vendure-plugin-goedgepickt/test/goedgepickt.spec.ts
@@ -1,13 +1,7 @@
-import {
-  createTestEnvironment,
-  registerInitializer,
-  SimpleGraphQLClient,
-  SqljsInitializer,
-  testConfig,
-} from '@vendure/testing';
-import { initialData } from '../../test/src/initial-data';
+import nock from 'nock';
 import {
   DefaultLogger,
+  ID,
   LanguageCode,
   LogLevel,
   mergeConfig,
@@ -19,41 +13,42 @@ import {
   ShippingMethodService,
   StockLevelService,
   TransactionalConnection,
-  ID,
 } from '@vendure/core';
-import { TestServer } from '@vendure/testing/lib/test-server';
 import {
-  GoedgepicktConfig,
+  createTestEnvironment,
+  registerInitializer,
+  SimpleGraphQLClient,
+  SqljsInitializer,
+  testConfig,
+} from '@vendure/testing';
+import { TestServer } from '@vendure/testing/lib/test-server';
+import gql from 'graphql-tag';
+import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import { getOrder } from '../../test/src/admin-utils';
+import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
+import { initialData } from '../../test/src/initial-data';
+import {
+  addItem,
+  createSettledOrder,
+  SettledOrder,
+} from '../../test/src/shop-utils';
+import { testPaymentMethod } from '../../test/src/test-payment-method';
+import {
   goedgepicktHandler,
   GoedgepicktPlugin,
   IncomingOrderStatusEvent,
   IncomingStockUpdateEvent,
   OrderInput,
 } from '../src';
-import nock from 'nock';
-import { GoedgepicktService } from '../src/api/goedgepickt.service';
-import {
-  getGoedgepicktConfig,
-  runGoedgepicktFullSync,
-  updateGoedgepicktConfig,
-} from '../src/ui/queries.graphql';
-import fs from 'fs';
-import path from 'path';
-import { compileUiExtensions } from '@vendure/ui-devkit/compiler';
 import { GoedgepicktClient } from '../src/api/goedgepickt.client';
-import { getOrder } from '../../test/src/admin-utils';
-import { addItem, createSettledOrder } from '../../test/src/shop-utils';
-import { testPaymentMethod } from '../../test/src/test-payment-method';
-import gql from 'graphql-tag';
-import { expect, describe, beforeAll, afterAll, it, vi, test } from 'vitest';
-import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
+import { GoedgepicktService } from '../src/api/goedgepickt.service';
+import { updateChannel } from './helpers';
 describe('Goedgepickt plugin', function () {
   const defaultChannelToken = 'e2e-default-channel';
   let server: TestServer;
   let adminClient: SimpleGraphQLClient;
   let shopClient: SimpleGraphQLClient;
   let serverStarted = false;
-  let productVariantService: ProductVariantService;
   const ggConfig = {
     apiKey: 'test-api-key',
     webshopUuid: 'test-webshop-uuid',
@@ -62,7 +57,7 @@ describe('Goedgepickt plugin', function () {
   let pushProductsPayloads: any[] = [];
   let createOrderPayload: OrderInput;
   let webhookPayloads: any[] = [];
-  let order: Order;
+  let order: SettledOrder;
   const apiUrl = 'https://account.goedgepickt.nl/';
   // Update products
   nock(apiUrl)
@@ -84,10 +79,12 @@ describe('Goedgepickt plugin', function () {
     });
   // Get order
   nock(apiUrl)
+    .persist(true)
     .get('/api/v1/orders/testUuid')
     .reply(200, {
       message: 'Order created',
       orderUuid: 'testUuid',
+      status: 'completed',
       shipments: [
         {
           trackTraceCode: 'XYZ',
@@ -97,13 +94,15 @@ describe('Goedgepickt plugin', function () {
     });
   // Find by SKU
   nock(apiUrl)
-    .persist(true)
     .get(
       /\/api\/v1\/products\?searchAttribute=sku&searchDelimiter=%3D&searchValue=*/
     )
-    .reply(200, { items: [] });
+    .times(3)
+    .reply(200, {
+      items: [],
+    });
   // Get webshops
-  nock(apiUrl)
+  nock('https://account.goedgepickt.nl')
     .persist(true)
     .get('/api/v1/webshops')
     .reply(200, { items: [{ uuid: ggConfig.webshopUuid }] });
@@ -152,41 +151,26 @@ describe('Goedgepickt plugin', function () {
     });
     serverStarted = true;
     await adminClient.asSuperAdmin();
-    productVariantService = server.app.get<ProductVariantService>(
-      ProductVariantService
-    );
   }, 60000);
 
   it('Should start successfully', async () => {
     await expect(serverStarted).toBe(true);
   });
 
-  it('Updates config via graphql and sets webhooks', async () => {
-    const result: { updateGoedgepicktConfig: GoedgepicktConfig } =
-      await adminClient.query(updateGoedgepicktConfig, {
-        input: ggConfig,
-      });
-    await expect(result.updateGoedgepicktConfig.webshopUuid).toBe(
-      ggConfig.webshopUuid
+  it('Updates credentials on Channel', async () => {
+    const uuidAndApiKey = ggConfig.webshopUuid + ':' + ggConfig.apiKey;
+    const result = await adminClient.query(updateChannel, {
+      input: {
+        id: 'T_1',
+        customFields: {
+          ggEnabled: true,
+          ggUuidApiKey: uuidAndApiKey,
+        },
+      },
+    });
+    await expect(result.updateChannel.customFields.ggUuidApiKey).toBe(
+      uuidAndApiKey
     );
-    await expect(result.updateGoedgepicktConfig.orderWebhookKey).toBe(
-      'test-secret'
-    );
-    await expect(result.updateGoedgepicktConfig.stockWebhookKey).toBe(
-      'test-secret'
-    );
-    await expect(webhookPayloads.length).toBe(2); // Order and Stock webhooks
-    await expect(webhookPayloads[0].targetUrl).toBe(
-      'https://test-host/goedgepickt/webhook/e2e-default-channel'
-    );
-  });
-
-  it('Retrieves config via graphql', async () => {
-    const result = await adminClient.query(getGoedgepicktConfig);
-    await expect(result.goedgepicktConfig.webshopUuid).toBe(
-      ggConfig.webshopUuid
-    );
-    await expect(result.goedgepicktConfig.apiKey).toBe(ggConfig.apiKey);
   });
 
   it('Pushes products and updates stocklevel on FullSync', async () => {
@@ -205,7 +189,13 @@ describe('Goedgepickt plugin', function () {
           },
         ],
       });
-    await adminClient.query(runGoedgepicktFullSync);
+    await adminClient.query(
+      gql`
+        mutation {
+          runGoedgepicktFullSync
+        }
+      `
+    );
     await new Promise((resolve) => setTimeout(resolve, 500)); // Some time for async event handling
     await expect(pushProductsPayloads.length).toBeGreaterThanOrEqual(3); // After multiple restarts we have 1 extra
     const laptopPayload = pushProductsPayloads.find(
@@ -271,34 +261,6 @@ describe('Goedgepickt plugin', function () {
     await expect(createOrderPayload.shippingMethod).toBe('Standard Shipping');
   });
 
-  it('Fails webhook with invalid signature', async () => {
-    const body: IncomingOrderStatusEvent = {
-      newStatus: 'completed',
-      orderNumber: order.code,
-      event: 'orderStatusChanged',
-      orderUuid: 'testUuid',
-    };
-    const signature = 'wrong-signature';
-    const res = await shopClient.fetch(
-      `http://localhost:3105/goedgepickt/webhook/${defaultChannelToken}`,
-      {
-        method: 'POST',
-        body: JSON.stringify(body),
-        headers: {
-          signature: signature,
-        },
-      }
-    );
-    const ctx = await server.app
-      .get(GoedgepicktService)
-      .getCtxForChannel(defaultChannelToken);
-    order = (await server.app
-      .get(OrderService)
-      .findOneByCode(ctx, order.code))!;
-    expect(res.ok).toBe(true);
-    expect(order.state).toBe('PaymentSettled');
-  });
-
   it('Completes order via webhook', async () => {
     const body: IncomingOrderStatusEvent = {
       newStatus: 'completed',
@@ -306,28 +268,24 @@ describe('Goedgepickt plugin', function () {
       event: 'orderStatusChanged',
       orderUuid: 'testUuid',
     };
-    const signature = GoedgepicktClient.computeSignature(
-      'test-secret',
-      JSON.stringify(body)
-    );
     const res = await shopClient.fetch(
       `http://localhost:3105/goedgepickt/webhook/${defaultChannelToken}`,
       {
         method: 'POST',
         body: JSON.stringify(body),
         headers: {
-          signature: signature,
+          signature: 'test-signature',
         },
       }
     );
     const ctx = await server.app
       .get(GoedgepicktService)
       .getCtxForChannel(defaultChannelToken);
-    order = (await server.app
+    const adminOrder = (await server.app
       .get(OrderService)
       .findOneByCode(ctx, order.code))!;
     expect(res.ok).toBe(true);
-    expect(order.state).toBe('Delivered');
+    expect(adminOrder.state).toBe('Delivered');
   });
 
   it('Has fulfillment after completion', async () => {
@@ -335,36 +293,6 @@ describe('Goedgepickt plugin', function () {
     const fulfillment = adminOrder?.fulfillments?.[0];
     expect(fulfillment?.method).toBe('GoedGepickt - pinelab.studio/xyz');
     expect(fulfillment?.trackingCode).toBe('XYZ');
-  });
-
-  it('Decreases stock via webhook', async () => {
-    const body: IncomingStockUpdateEvent = {
-      event: 'stockUpdated',
-      newStock: '123',
-      productSku: 'L2201308',
-      productUuid: 'doesntmatter',
-    };
-    const signature = GoedgepicktClient.computeSignature(
-      'test-secret',
-      JSON.stringify(body)
-    );
-    const res = await shopClient.fetch(
-      `http://localhost:3105/goedgepickt/webhook/${defaultChannelToken}`,
-      {
-        method: 'POST',
-        body: JSON.stringify(body),
-        headers: {
-          signature: signature,
-        },
-      }
-    );
-    // Await async job processing
-    await new Promise((resolve) => setTimeout(resolve, 200));
-    const updatedVariant = await findVariantBySku('L2201308');
-    expect(res.ok).toBe(true);
-    expect(updatedVariant).toBeDefined();
-    const stockOnHand = await getAvailableStock(updatedVariant?.id!);
-    expect(stockOnHand).toBe(123);
   });
 
   it('Pushes product on product creation', async () => {
@@ -397,6 +325,47 @@ describe('Goedgepickt plugin', function () {
     await new Promise((resolve) => setTimeout(resolve, 500)); // Some time for async event handling
     const payload = pushProductsPayloads.find((p) => p.sku === 'sku123');
     expect(payload).toBeDefined();
+  });
+
+  it('Decreases stock via webhook', async () => {
+    // Find by SKU
+    nock.cleanAll();
+    nock(apiUrl)
+      .get(
+        /\/api\/v1\/products\?searchAttribute=sku&searchDelimiter=%3D&searchValue=*/
+      )
+      .reply(200, {
+        items: [
+          {
+            uuid: 'test-uuid',
+            sku: 'L2201316',
+            stock: {
+              freeStock: 123,
+            },
+          },
+        ],
+      });
+    const body: IncomingStockUpdateEvent = {
+      event: 'stockUpdated',
+      newStock: 'doestn matter',
+      productSku: 'L2201316',
+      productUuid: 'doesntmatter',
+    };
+    const res = await shopClient.fetch(
+      `http://localhost:3105/goedgepickt/webhook/${defaultChannelToken}`,
+      {
+        method: 'POST',
+        body: JSON.stringify(body),
+        headers: {
+          signature: 'some-sample-signature',
+        },
+      }
+    );
+    const updatedVariant = await findVariantBySku('L2201316');
+    expect(res.ok).toBe(true);
+    expect(updatedVariant).toBeDefined();
+    const stockOnHand = await getAvailableStock(updatedVariant?.id!);
+    expect(stockOnHand).toBe(123);
   });
 
   if (process.env.TEST_ADMIN_UI) {