+++ packages/vendure-plugin-goedgepickt/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.0.0 (2025-03-20)
+
+- Removed validating of webhook signature. Fetching stock and order status directly from GoedGepickt on incoming webhook instead.
+- Moved GG config to Channel Custom Fields, instead of custom UI component.
+- Breaking: DB migration needed
+- Breaking: You need to save GoedGepickt credentials on a channel now: Settings > Channel > GoedGepickt custom fields
+
 # 1.4.1 (2025-03-18)
 
 - Only sync product image on creation, not on update, after complaints from GoedGepickt about too many updates.